requirements.txt
autopep8==1.4.4
beautifulsoup4==4.6.1
certifi==2018.11.29
chardet==3.0.4
Django==2.2
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.1.0
django-multiselectfield==0.1.8
djangorestframework==3.9.3
html5lib==1.0.1
idna==2.7
Markdown==3.1
onesignal-notifications==0.2.2
pep8==1.7.1
Pillow==5.3.0
pycodestyle==2.5.0
pytz==2018.5
requests==2.21.0
six==1.11.0
sqlparse==0.3.0
tika==1.19
urllib3==1.24.2
webencodings==0.5.1
